+++ public-info/src/main/java/com/bank/publicinfo/dto/BranchDto.java
@@ -4,7 +4,7 @@ import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.Set;
 
-public record BranchDto(BigInteger id,
+public record BranchDto(Long id,
                         String address,
                         BigInteger phoneNumber,
                         String city,
