+++ public-info/src/main/java/com/bank/publicinfo/repository/LicenseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface LicenseRepository extends JpaRepository<License, BigInteger> {
+public interface LicenseRepository extends JpaRepository<License, Long> {
 }
